+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @authhero/demo
 
+## 0.7.7
+
+### Patch Changes
+
+- Updated dependencies
+  - authhero@0.45.0
+
+## 0.7.6
+
+### Patch Changes
+
+- Updated dependencies
+  - authhero@0.44.0
+
+## 0.7.5
+
+### Patch Changes
+
+- Updated dependencies
+  - authhero@0.43.0
+
 ## 0.7.4
 
 ### Patch Changes